pyproject.toml
[tool.poetry]
name = "art-deco"
version = "0.1.0"
description = ""
authors = ["baha <baktiarzaksylykov22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pydantic = "^2.8.2"
aiosqlite = "^0.20.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
alembic = "^1.13.2"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
black = "^24.4.2"
fastapi-pagination = "^0.12.26"
sqladmin = "^0.18.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
